requirements.txt
fastapi[standard]
uvicorn[standard]
pymongo
python-dotenv
python-multipart
cloudinary
bcrypt